% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azcopy.R
\name{call_azcopy}
\alias{call_azcopy}
\alias{azcopy_login}
\title{Call the azcopy file transfer utility}
\usage{
call_azcopy(...)

azcopy_login(force = FALSE)
}
\arguments{
\item{...}{Arguments to pass to azcopy. If no arguments are supplied, azcopy will print a help screen.}

\item{force}{For \code{azcopy_login}, whether to force azcopy to relogin.}
}
\description{
Call the azcopy file transfer utility
}
