% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_client_funcs.R
\name{list_blobs}
\alias{list_blobs}
\alias{upload_blob}
\alias{download_blob}
\alias{delete_blob}
\title{Operations on a blob container}
\usage{
list_blobs(container, info = c("partial", "name", "all"))

upload_blob(container, src, dest, type = "BlockBlob")

download_blob(container, src, dest, overwrite = FALSE)

delete_blob(container, blob, confirm = TRUE)
}
\arguments{
\item{container}{A blob container object.}

\item{info}{For \code{list_blobs}, level of detail about each blob to return: a vector of names only; the name, size and last-modified date (default); or all information.}

\item{src, dest}{The source and destination filenames for uploading and downloading. Paths are allowed.}

\item{blob}{A string naming a blob.}

\item{confirm}{Whether to ask for confirmation on deleting a blob.}
}
\value{
For \code{list_blobs}, details on the blobs in the container.
}
\description{
Upload, download, or delete a blob; list blobs in a container.
}
\seealso{
\link{blob_container}, \link{az_storage}
}
