% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{download_from_url}
\alias{download_from_url}
\alias{upload_to_url}
\title{Generic upload and download}
\usage{
download_from_url(src, dest, ..., overwrite = FALSE)

upload_to_url(src, dest, ...)
}
\arguments{
\item{src, }{dest The source and destination files/URLs. Paths are allowed.}

\item{...}{Further arguments to pass to lower-level functions. In particular, use \code{key} and/or \code{sas} to supply an access key or SAS. Without a key or SAS, only unauthenticated (anonymous) access is possible.}

\item{overwrite}{For downloading, whether to overwrite any destination files that exist.}
}
\value{
NULL on successful completion.
}
\description{
Generic upload and download
}
\details{
These functions allow you to transfer files to and from a storage account, given the URL of the destination (for uploading) or source (for downloading). They dispatch to \link{upload_azure_file}/\link{download_azure_file} for a file storage URL and \link{upload_blob}/\link{download_blob} for a blob storage URL respectively.
}
\seealso{
\link{download_azure_file}, \link{download_blob}. \link{az_storage}
}
