% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_properties.R
\name{get_storage_properties}
\alias{get_storage_properties}
\alias{get_storage_properties.blob_endpoint}
\alias{get_storage_properties.file_endpoint}
\alias{get_storage_properties.blob_container}
\alias{get_storage_properties.file_share}
\alias{get_storage_properties.adls_filesystem}
\alias{get_adls_file_acl}
\alias{get_adls_file_status}
\title{Get storage properties for an object}
\usage{
get_storage_properties(object, ...)

\method{get_storage_properties}{blob_endpoint}(object, ...)

\method{get_storage_properties}{file_endpoint}(object, ...)

\method{get_storage_properties}{blob_container}(object, blob, ...)

\method{get_storage_properties}{file_share}(object, file, isdir, ...)

\method{get_storage_properties}{adls_filesystem}(object, file, ...)

get_adls_file_acl(filesystem, file)

get_adls_file_status(filesystem, file)
}
\arguments{
\item{object}{A storage object: an endpoint, blob container, file share, or ADLS filesystem.}

\item{file}{The name of an individual blob, file or directory within a container.}

\item{isdir}{Whether the object is a file or directory.}

\item{filesystem}{An ADLS filesystem.}
}
\value{
A list describing the object properties.
}
\description{
Get storage properties for an object
}
\details{
The \code{get_storage_properties} generic returns a list of properties for the given storage object. There are methods defined for objects of class \code{storage_endpoint}, \code{blob_container}, \code{file_share} and \code{adls_filesystem}. Similar functions are defined for individual blobs, files and directories.
}
\seealso{
\link{storage_endpoint}, \link{blob_container}, \link{file_share}

[Blob service properties reference[(https://docs.microsoft.com/en-us/rest/api/storageservices/get-blob-service-properties).
\href{https://docs.microsoft.com/en-us/rest/api/storageservices/get-file-service-properties}{File service properties reference},
\href{https://docs.microsoft.com/en-us/rest/api/storageservices/get-container-properties}{Blob container properties reference},
\href{https://docs.microsoft.com/en-us/rest/api/storageservices/get-share-properties}{File share properties reference},
\href{https://docs.microsoft.com/en-us/rest/api/storageservices/datalakestoragegen2/filesystem/getproperties}{ADLS filesystem properties reference}
}
