% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_request.R
\name{sign_request}
\alias{sign_request}
\title{Signs a request to the storage REST endpoint with a shared key}
\usage{
sign_request(endpoint, ...)
}
\arguments{
\item{endpoint}{An endpoint object.}

\item{...}{Further arguments to pass to individual methods.}
}
\value{
A named list of request headers. One of these should be the \code{Authorization} header containing the request signature.
}
\description{
Signs a request to the storage REST endpoint with a shared key
}
\details{
This is a generic method to allow for variations in how the different storage services handle key authorisation. The default method works with blob, file and ADLSgen2 storage.
}
