% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_utils.R
\name{do_container_op}
\alias{do_container_op}
\alias{do_storage_call}
\title{Call a storage account endpoint}
\usage{
do_container_op(container, path = "", options = list(),
  headers = list(), http_verb = "GET", ...)

do_storage_call(endpoint_url, path, options = list(), headers = list(),
  body = NULL, ..., key = NULL, token = NULL, sas = NULL,
  api_version = getOption("azure_storage_api_version"),
  http_verb = c("GET", "DELETE", "PUT", "POST", "HEAD", "PATCH"),
  http_status_handler = c("stop", "warn", "message", "pass"),
  progress = NULL)
}
\description{
Call a storage account endpoint
}
