% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{storage_endpoint}
\alias{storage_endpoint}
\alias{endpoint}
\alias{blob_endpoint}
\alias{file_endpoint}
\alias{queue_endpoint}
\alias{table_endpoint}
\title{Create a storage endpoint object}
\usage{
storage_endpoint(endpoint, key = NULL, sas = NULL,
  api_version = getOption("azure_storage_api_version"))
}
\arguments{
\item{endpoint}{The URL (hostname) for the endpoint. This must be of the form \code{http[s]://{account-name}.{type}.{core-host-name}}, where \code{type} is one of \code{"blob"}, \code{"file"}, \code{"queue"} or \code{"table"}. On the public Azure cloud, endpoints will be of the form \code{https://{account-name}.{type}.core.windows.net}.}

\item{key}{The access key for the storage account.}

\item{sas}{A shared access signature for the account. If neither \code{key} nor \code{sas} are provided, only public (anonymous) access to the endpoint is possible.}

\item{api_version}{The storage API version to use when interacting with the host. Currently defaults to \code{"2017-07-29"}.}
}
\value{
An object of S3 class \code{"blob_endpoint"}, \code{"file_endpoint"}, \code{"queue_endpoint"} or \code{"table_endpoint"} depending on the type of endpoint. All of these also inherit from class \code{"storage_endpoint"}.
}
\description{
Create a storage endpoint object
}
\details{
This is the starting point for the client-side storage interface in AzureRMR.
}
\seealso{
\link{az_storage}, \link{file_share}, \link{create_file_share}, \link{blob_container}, \link{create_blob_container}
}
