% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_methods.R
\name{delete_storage_account}
\alias{delete_storage_account}
\title{Delete an Azure storage account}
\usage{
delete_storage_account(name, confirm=TRUE, wait=FALSE)
}
\arguments{
\item{name}{The name of the storage account.}

\item{confirm}{Whether to ask for confirmation before deleting.}

\item{wait}{Whether to wait until the deletion is complete.}
}
\value{
NULL on successful deletion.
}
\description{
Method for the \link[AzureRMR:az_resource_group]{AzureRMR::az_resource_group} class.
}
\seealso{
\link{create_storage_account}, \link{get_storage_account}, \link{az_storage},
\href{https://docs.microsoft.com/en-us/rest/api/storagerp/}{Azure Storage Provider API reference}
}
